archive/sokuk_legacy/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robust-streamlit-app"
version = "1.0.0"
description = "Maximal robuste Streamlit-Anwendung"
authors = [{name = "Developer", email = "dev@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core
    "streamlit==1.28.1",
    "pydantic==2.5.0",
    "sqlalchemy==2.0.23",
    "alembic==1.12.1",
    "duckdb==0.9.2",
    "pandas==2.1.3",
    "numpy==1.25.2",
    
    # Jobs & Async
    "apscheduler==3.10.4",
    "redis==5.0.1",
    
    # Security
    "python-dotenv==1.0.0",
    "cryptography==41.0.7",
    "bcrypt==4.1.1",
    
    # CLI
    "typer==0.9.0",
    "rich==13.7.0",
    
    # Metrics & Monitoring
    "prometheus-client==0.19.0",
    "structlog==23.2.0",
    
    # Utils
    "python-multipart==0.0.6",
    "httpx==0.25.2",
    "tenacity==8.2.3",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "hypothesis==6.88.1",
    "playwright==1.40.0",
    
    # Linting & Formatting
    "ruff==0.1.6",
    "black==23.11.0",
    "mypy==1.7.1",
    "pre-commit==3.5.0",
    
    # Database
    "psycopg2-binary==2.9.9",
]

prod = [
    "psycopg2-binary==2.9.9",
    "gunicorn==21.2.0",
]

[project.scripts]
app = "cli.main:app"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "PIE", "SIM", "RET"]
ignore = ["E501", "B008"]

[tool.black]
target-version = ["py311"]
line-length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=core --cov=models --cov=repositories --cov=services --cov-report=html --cov-report=term-missing"